+++ include/socket_manager/recv_waker.h
@@ -14,11 +14,11 @@ public:
   void operator=(const RcvWaker &) = delete;
 
 private:
-  explicit RcvWaker(CWaker* waker);
+  explicit RcvWaker(CWaker *waker);
 
-  friend int ::socket_manager_extern_on_msg(struct OnMsgObj this_, ConnMsg msg, CWaker *waker, char **err);
+  friend long::socket_manager_extern_on_msg(struct OnMsgObj this_, ConnMsg msg, CWaker *waker, char **err);
 
-  CWaker* waker;
+  CWaker *waker;
 };
 
 #endif //SOCKET_MANAGER_RECV_WAKER_H
